+++ adminer/include/coverage.inc.php
@@ -15,6 +15,8 @@ if (extension_loaded("xdebug") && file_exists(sys_get_temp_dir() . "/adminer.cov
 			file_put_contents($coverage_filename, serialize($coverage));
 		}
 	}
+	// Xdebug marks the not executed lines when it compiles a file, OPcache serves the cached files without compiling them
+	\ini_set("opcache.enable", '0'); // PHP allows only disabling this setting at runtime; Adminer\ini_set() is not defined yet
 	xdebug_start_code_coverage(XDEBUG_CC_UNUSED | XDEBUG_CC_DEAD_CODE);
 	register_shutdown_function('Adminer\save_coverage');
 }