+++ CHANGELOG.md
@@ -135,6 +135,7 @@
 - Rename the Git master branch to main
 - Run the command line tests by composer test and in CI
 - Convert the end-to-end tests from Katalon Recorder to Playwright, run them headless by composer e2e
+- Fix the code coverage reporting all files as fully covered when OPcache is enabled
 
 ## Adminer 5.5.1 (released 2026-07-21)
 - Ignore invalid X-Forwarded-Prefix (GHSA-fr74-9mf9-gf44)